Refactor code to allow for custom (user-defined) timing functions
This allows embedded implementations/interfaces to add their own timer
functions, and initialization/teardown code for the timers.
