+++ .github/workflows/embedded.yaml
@@ -48,7 +48,7 @@ jobs:
                   -DEMBEDDED=${{ matrix.embedded }} \
                   -DALGEBRA='default' \
                   -DLONG=OFF \
-                  -DPROFILING=OFF \
+                  -DOSQP_ENABLE_PROFILING=OFF \
                   -DOSQP_ENABLE_INTERRUPT=OFF \
                   -DOSQP_ENABLE_PRINTING=OFF \
                   -DOSQP_BUILD_UNITTESTS=OFF
