+++ include/private/util.h
@@ -67,81 +67,6 @@ void print_footer(OSQPInfo *info,
 # endif /* ifdef OSQP_ENABLE_PRINTING */
 
 
-/*********************************
-* Timer Structs and Functions * *
-*********************************/
-
-/*! \cond PRIVATE */
-
-# ifdef PROFILING
-
-// Windows
-#  ifdef IS_WINDOWS
-
-  // Some R packages clash with elements
-  // of the windows.h header, so use a
-  // slimmer version for conflict avoidance
-# ifdef R_LANG
-#define NOGDI
-# endif
-
-#   include <windows.h>
-
-struct OSQP_TIMER {
-  LARGE_INTEGER tic;
-  LARGE_INTEGER toc;
-  LARGE_INTEGER freq;
-};
-
-// Mac
-#  elif defined IS_MAC
-
-#   include <mach/mach_time.h>
-
-/* Use MAC OSX  mach_time for timing */
-struct OSQP_TIMER {
-  uint64_t                  tic;
-  uint64_t                  toc;
-  mach_timebase_info_data_t tinfo;
-};
-
-// Linux
-#  else /* ifdef IS_WINDOWS */
-
-/* Use POSIX clock_gettime() for timing on non-Windows machines */
-#   include <time.h>
-#   include <sys/time.h>
-
-
-struct OSQP_TIMER {
-  struct timespec tic;
-  struct timespec toc;
-};
-
-#  endif /* ifdef IS_WINDOWS */
-
-/*! \endcond */
-
-/**
- * Timer Methods
- */
-
-/**
- * Start timer
- * @param t Timer object
- */
-void    osqp_tic(OSQPTimer *t);
-
-/**
- * Report time
- * @param  t Timer object
- * @return   Reported time
- */
-c_float osqp_toc(OSQPTimer *t);
-
-# endif /* END #ifdef PROFILING */
-
-
 /* ================================= DEBUG FUNCTIONS ======================= */
 
 /*! \cond PRIVATE */
