+++ src/osqp_api.c
@@ -8,6 +8,8 @@
 #include "version.h"
 #include "lin_alg.h"
 #include "printing.h"
+#include "timing.h"
+
 
 #ifdef OSQP_CODEGEN
   #include "codegen.h"
@@ -148,11 +150,11 @@ c_int osqp_setup(OSQPSolver         **solverp,
   if (!(solver->info)) return osqp_error(OSQP_MEM_ALLOC_ERROR);
 
   // Start and allocate directly timer
-# ifdef PROFILING
-  work->timer = c_malloc(sizeof(OSQPTimer));
+# ifdef OSQP_ENABLE_PROFILING
+  work->timer = OSQPTimer_new();
   if (!(work->timer)) return osqp_error(OSQP_MEM_ALLOC_ERROR);
   osqp_tic(work->timer);
-# endif /* ifdef PROFILING */
+# endif /* ifdef OSQP_ENABLE_PROFILING */
 
   // Initialize algebra libraries
   exitflag = osqp_algebra_init_libs(settings->device);
@@ -313,7 +315,7 @@ c_int osqp_setup(OSQPSolver         **solverp,
   // Initialize information
   solver->info->status_polish = 0;              // Polishing not performed
   update_status(solver->info, OSQP_UNSOLVED);
-# ifdef PROFILING
+# ifdef OSQP_ENABLE_PROFILING
   solver->info->solve_time  = 0.0;                   // Solve time to zero
   solver->info->update_time = 0.0;                   // Update time to zero
   solver->info->polish_time = 0.0;                   // Polish time to zero
@@ -323,7 +325,7 @@ c_int osqp_setup(OSQPSolver         **solverp,
   work->first_run         = 1;
   work->clear_update_time = 0;
   work->rho_update_from_solve = 0;
-# endif /* ifdef PROFILING */
+# endif /* ifdef OSQP_ENABLE_PROFILING */
   solver->info->rho_updates  = 0;                      // Rho updates set to 0
   solver->info->rho_estimate = solver->settings->rho;  // Best rho estimate
   solver->info->obj_val      = OSQP_INFTY;
@@ -339,7 +341,7 @@ c_int osqp_setup(OSQPSolver         **solverp,
 
   // If adaptive rho and automatic interval, but profiling disabled, we need to
   // set the interval to a default value
-# ifndef PROFILING
+# ifndef OSQP_ENABLE_PROFILING
   if (solver->settings->adaptive_rho && !solver->settings->adaptive_rho_interval) {
     if (solver->settings->check_termination) {
       // If check_termination is enabled, we set it to a multiple of the check
@@ -351,7 +353,7 @@ c_int osqp_setup(OSQPSolver         **solverp,
       solver->settings->adaptive_rho_interval = OSQP_ADAPTIVE_RHO_FIXED;
     }
   }
-# endif /* ifndef PROFILING */
+# endif /* ifndef OSQP_ENABLE_PROFILING */
 
   // Return exit flag
   return 0;
@@ -368,9 +370,9 @@ c_int osqp_solve(OSQPSolver *solver) {
   c_int can_check_termination; // boolean: check termination or not
   OSQPWorkspace* work;
 
-#ifdef PROFILING
+#ifdef OSQP_ENABLE_PROFILING
   c_float temp_run_time;       // Temporary variable to store current run time
-#endif /* ifdef PROFILING */
+#endif /* ifdef OSQP_ENABLE_PROFILING */
 
 #ifdef OSQP_ENABLE_PRINTING
   c_int can_print;             // Boolean whether you can print
@@ -380,11 +382,11 @@ c_int osqp_solve(OSQPSolver *solver) {
   if (!solver || !solver->work) return osqp_error(OSQP_WORKSPACE_NOT_INIT_ERROR);
   work = solver->work;
 
-#ifdef PROFILING
+#ifdef OSQP_ENABLE_PROFILING
   if (work->clear_update_time == 1)
     solver->info->update_time = 0.0;
   work->rho_update_from_solve = 1;
-#endif /* ifdef PROFILING */
+#endif /* ifdef OSQP_ENABLE_PROFILING */
 
   // Initialize variables
   exitflag              = 0;
@@ -397,9 +399,9 @@ c_int osqp_solve(OSQPSolver *solver) {
   compute_obj = 0;
 #endif /* ifdef OSQP_ENABLE_PRINTING */
 
-#ifdef PROFILING
+#ifdef OSQP_ENABLE_PROFILING
   osqp_tic(work->timer); // Start timer
-#endif /* ifdef PROFILING */
+#endif /* ifdef OSQP_ENABLE_PROFILING */
 
 
 #ifdef OSQP_ENABLE_PRINTING
@@ -454,7 +456,7 @@ c_int osqp_solve(OSQPSolver *solver) {
     }
 #endif /* ifdef OSQP_ENABLE_INTERRUPT */
 
-#ifdef PROFILING
+#ifdef OSQP_ENABLE_PROFILING
 
     // Check if solver time_limit is enabled. In case, check if the current
     // run time is more than the time_limit option.
@@ -475,7 +477,7 @@ c_int osqp_solve(OSQPSolver *solver) {
 # endif /* ifdef OSQP_ENABLE_PRINTING */
       break;
     }
-#endif /* ifdef PROFILING */
+#endif /* ifdef OSQP_ENABLE_PROFILING */
 
 
     // Can we check for termination ?
@@ -524,7 +526,7 @@ c_int osqp_solve(OSQPSolver *solver) {
 
 
 #if EMBEDDED != 1
-# ifdef PROFILING
+# ifdef OSQP_ENABLE_PROFILING
 
     // If adaptive rho with automatic interval, check if the solve time is a
     // certain fraction
@@ -556,7 +558,7 @@ c_int osqp_solve(OSQPSolver *solver) {
           solver->settings->check_termination);
       } // If time condition is met
     }   // If adaptive rho enabled and interval set to auto®
-# else // PROFILING
+# else // OSQP_ENABLE_PROFILING
     if (solver->settings->adaptive_rho && !solver->settings->adaptive_rho_interval) {
       // Set adaptive_rho_interval to constant value
       if (solver->settings->check_termination) {
@@ -569,7 +571,7 @@ c_int osqp_solve(OSQPSolver *solver) {
         solver->settings->adaptive_rho_interval = OSQP_ADAPTIVE_RHO_FIXED;
       }
     }
-# endif // PROFILING
+# endif // OSQP_ENABLE_PROFILING
 
     // Adapt rho
     if (solver->settings->adaptive_rho &&
@@ -652,14 +654,14 @@ c_int osqp_solve(OSQPSolver *solver) {
     }
   }
 
-#ifdef PROFILING
+#ifdef OSQP_ENABLE_PROFILING
   /* if time-limit reached check termination and update status accordingly */
  if (solver->info->status_val == OSQP_TIME_LIMIT_REACHED) {
     if (!check_termination(solver, 1)) { // Try for approximate solutions
       update_status(solver->info, OSQP_TIME_LIMIT_REACHED); /* Change update status back to OSQP_TIME_LIMIT_REACHED */
     }
   }
-#endif /* ifdef PROFILING */
+#endif /* ifdef OSQP_ENABLE_PROFILING */
 
 
 #if EMBEDDED != 1
@@ -668,9 +670,9 @@ c_int osqp_solve(OSQPSolver *solver) {
 #endif /* if EMBEDDED != 1 */
 
   /* Update solve time */
-#ifdef PROFILING
+#ifdef OSQP_ENABLE_PROFILING
   solver->info->solve_time = osqp_toc(work->timer);
-#endif /* ifdef PROFILING */
+#endif /* ifdef OSQP_ENABLE_PROFILING */
 
 
 #ifndef EMBEDDED
@@ -679,7 +681,7 @@ c_int osqp_solve(OSQPSolver *solver) {
     polish(solver);
 #endif /* ifndef EMBEDDED */
 
-#ifdef PROFILING
+#ifdef OSQP_ENABLE_PROFILING
   /* Update total time */
   if (work->first_run) {
     // total time: setup + solve + polish
@@ -701,7 +703,7 @@ c_int osqp_solve(OSQPSolver *solver) {
 
   // Indicate that osqp_update_rho is not called from osqp_solve
   work->rho_update_from_solve = 0;
-#endif /* ifdef PROFILING */
+#endif /* ifdef OSQP_ENABLE_PROFILING */
 
 #ifdef OSQP_ENABLE_PRINTING
   /* Print final footer */
@@ -713,9 +715,9 @@ c_int osqp_solve(OSQPSolver *solver) {
 
 
 // Define exit flag for quitting function
-#if defined(PROFILING) || defined(OSQP_ENABLE_INTERRUPT) || EMBEDDED != 1
+#if defined(OSQP_ENABLE_PROFILING) || defined(OSQP_ENABLE_INTERRUPT) || EMBEDDED != 1
 exit:
-#endif /* if defined(PROFILING) || defined(OSQP_ENABLE_INTERRUPT) || EMBEDDED != 1 */
+#endif /* if defined(OSQP_ENABLE_PROFILING) || defined(OSQP_ENABLE_INTERRUPT) || EMBEDDED != 1 */
 
 #ifdef OSQP_ENABLE_INTERRUPT
   // Restore previous signal handler
@@ -821,10 +823,10 @@ c_int osqp_cleanup(OSQPSolver *solver) {
     // Free information
     if (solver->info) c_free(solver->info);
 
-# ifdef PROFILING
+# ifdef OSQP_ENABLE_PROFILING
     // Free timer
-    if (work->timer) c_free(work->timer);
-# endif /* ifdef PROFILING */
+    if (work->timer) OSQPTimer_free(work->timer);
+# endif /* ifdef OSQP_ENABLE_PROFILING */
 
     // Free work
     c_free(work);
@@ -857,14 +859,14 @@ c_int osqp_update_data_vec(OSQPSolver    *solver,
   if (!solver || !solver->work) return osqp_error(OSQP_WORKSPACE_NOT_INIT_ERROR);
   work = solver->work;
 
-#ifdef PROFILING
+#ifdef OSQP_ENABLE_PROFILING
   if (work->clear_update_time == 1) {
     work->clear_update_time = 0;
     solver->info->update_time = 0.0;
   }
   /* Start timer */
   osqp_tic(work->timer);
-#endif /* ifdef PROFILING */
+#endif /* ifdef OSQP_ENABLE_PROFILING */
 
   /* Update constraint bounds */
   if (l_new || u_new) {
@@ -910,9 +912,9 @@ c_int osqp_update_data_vec(OSQPSolver    *solver,
   /* Reset solver information */
   reset_info(solver->info);
 
-#ifdef PROFILING
+#ifdef OSQP_ENABLE_PROFILING
   solver->info->update_time += osqp_toc(work->timer);
-#endif /* ifdef PROFILING */
+#endif /* ifdef OSQP_ENABLE_PROFILING */
 
   return exitflag;
 }
@@ -983,13 +985,13 @@ c_int osqp_update_data_mat(OSQPSolver    *solver,
   if (!solver || !solver->work) return osqp_error(OSQP_WORKSPACE_NOT_INIT_ERROR);
   work = solver->work;
 
-#ifdef PROFILING
+#ifdef OSQP_ENABLE_PROFILING
   if (work->clear_update_time == 1) {
     work->clear_update_time = 0;
     solver->info->update_time = 0.0;
   }
   osqp_tic(work->timer); // Start timer
-#endif /* ifdef PROFILING */
+#endif /* ifdef OSQP_ENABLE_PROFILING */
 
   nnzP = OSQPMatrix_get_nz(work->data->P);
   nnzA = OSQPMatrix_get_nz(work->data->A);
@@ -1052,9 +1054,9 @@ c_int osqp_update_data_mat(OSQPSolver    *solver,
 
   if (exitflag != 0){c_eprint("new KKT matrix is not quasidefinite");}
 
-#ifdef PROFILING
+#ifdef OSQP_ENABLE_PROFILING
   solver->info->update_time += osqp_toc(work->timer);
-#endif /* ifdef PROFILING */
+#endif /* ifdef OSQP_ENABLE_PROFILING */
 
   return exitflag;
 }
@@ -1076,7 +1078,7 @@ c_int osqp_update_rho(OSQPSolver *solver,
     return 1;
   }
 
-#ifdef PROFILING
+#ifdef OSQP_ENABLE_PROFILING
   if (work->rho_update_from_solve == 0) {
     if (work->clear_update_time == 1) {
       work->clear_update_time = 0;
@@ -1084,7 +1086,7 @@ c_int osqp_update_rho(OSQPSolver *solver,
     }
     osqp_tic(work->timer); // Start timer
   }
-#endif /* ifdef PROFILING */
+#endif /* ifdef OSQP_ENABLE_PROFILING */
 
   // Update rho in settings
   solver->settings->rho = c_min(c_max(rho_new, OSQP_RHO_MIN), OSQP_RHO_MAX);
@@ -1106,10 +1108,10 @@ c_int osqp_update_rho(OSQPSolver *solver,
   // Update rho_vec in KKT matrix
   exitflag = work->linsys_solver->update_rho_vec(work->linsys_solver, work->rho_vec, solver->settings->rho);
 
-#ifdef PROFILING
+#ifdef OSQP_ENABLE_PROFILING
   if (work->rho_update_from_solve == 0)
     solver->info->update_time += osqp_toc(work->timer);
-#endif /* ifdef PROFILING */
+#endif /* ifdef OSQP_ENABLE_PROFILING */
 
   return exitflag;
 }
