+++ src/polish.c
@@ -5,6 +5,7 @@
 #include "lin_sys.h"
 #include "proj.h"
 #include "error.h"
+#include "timing.h"
 
 /**
  * Form reduced matrix A that contains only rows that are active at the
@@ -277,9 +278,9 @@ c_int polish(OSQPSolver *solver) {
   OSQPSettings*  settings  = solver->settings;
   OSQPWorkspace* work      = solver->work;
 
-#ifdef PROFILING
+#ifdef OSQP_ENABLE_PROFILING
   osqp_tic(work->timer); // Start timer
-#endif /* ifdef PROFILING */
+#endif /* ifdef OSQP_ENABLE_PROFILING */
 
   // Form Ared by assuming the active constraints and store in work->pol->Ared
   mred = form_Ared(work);
